// Issuance-dash-privqa/src/api/errorTypes.ts
export enum ErrorType {
  BAD_REQUEST = "BAD_REQUEST",
  FORBIDDEN = "FORBIDDEN",
  INCORRECT_CODE = "INCORRECT_CODE",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  INVALID_INPUT_FORMAT = "INVALID_INPUT_FORMAT",
  INVALID_PHONE_NUMBER = "INVALID_PHONE_NUMBER",
  PHONE_NUMBER_ALREADY_EXISTS = "PHONE_NUMBER_ALREADY_EXISTS",
  PHONE_NUMBER_DOES_NOT_EXISTS = "PHONE_NUMBER_ALREADY_VERIFIED",
  UNAUTHORIZED = "UNAUTHORIZED",
  UNKNOWN_TWILIO_ERROR = "UNKNOWN_TWILIO_ERROR",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  NOT_FOUND = "NOT_FOUND",
  UNVERIFIED_PHONE_NUMBER = "UNVERIFIED_PHONE_NUMBER",
  EMAIL_ALREADY_EXISTS = "EMAIL_ALREADY_EXISTS",
  FRAUDULENT_USER = "FRAUDULENT_USER",
  UNKNOWN_SEON_ERROR = "UNKNOWN_SEON_ERROR",
  EMAIL_DOES_NOT_EXISTS = "EMAIL_DOES_NOT_EXISTS",
  TRANSACTION_NOT_FOUND = "TRANSACTION_NOT_FOUND",
  SUBACCOUNT_NOT_FOUND = "SUBACCOUNT_NOT_FOUND",
  USERS_ALREADY_IN_DATABASE = "USERS_ALREADY_IN_DATABASE",
  MERCHANT_ALREADY_EXISTS = "MERCHANT_ALREADY_EXISTS",
  ISSUANCE_BANK_NOT_FOUND = "ISSUANCE_BANK_NOT_FOUND",
  INVALID_EMAIL = "INVALID_EMAIL",
  INVALID_ROLE = "INVALID_ROLE",
  INVITATION_NOT_FOUND = "INVITATION_NOT_FOUND",
  TOKEN_EXPIRED = "TOKEN_EXPIRED",
  EXTERNAL_SEND_INSIDE_THE_SYSTEM = "EXTERNAL_SEND_INSIDE_THE_SYSTEM",
  CONVERSION_RATE_ALREADY_ADDED = "CONVERSION_RATE_ALREADY_ADDED",
  WALLET_FEE_CONFIG_ALREADY_ADDED = "WALLET_FEE_CONFIG_ALREADY_ADDED",
  TRANSACTION_LIMIT_ALREADY_ADDED = "TRANSACTION_LIMIT_ALREADY_ADDED",
  NO_DATA_SENT = "NO_DATA_SENT",
  LANGUAGE_ALREADY_ADDED = "LANGUAGE_ALREADY_ADDED",
  LANGUAGE_MAPPED_WITH_INSTITUTIONS = "LANGUAGE_MAPPED_WITH_INSTITUTIONS"
}